Add support for multiple instances of CAC
User can now specify the number of CACs to deploy using the variable
cac_instance_count. Defaults to 1 if nothing specified.

CACs will be deployed to the same subnet.

To keep things simple, removed the option to specify a specific private
IP for the CACs.

The SSH output has also been removed - did not bother to figure out how
to output a mix of string (user name) and array (IP addresses).

Signed-off-by: Sherman Yin <syin@teradici.com>