Fixes CAC stuck waiting for AD service account
cloud-access-connector installation requires the AD service account it
uses to exist on the domain controller. The startup scripts ensures the
service account is available by looping on ldapwhoami against the
service account until it is successful before continuing with the
installation.

This has worked until the week of Jun 10, 2019, where it was noticed
that CAC installation is stuck in the loop of ldapwhoami. Debug logs
shows problems with resolving the domain name / domain controller, even
when a separate SSH session into the same VM running the same ldapwhoami
command at the same time is successful. The ldapwhoami loop eventually
exits succeedfully (e.g. name resolution finally works), but sometimes
after as long as 90 minutes.

This commit applies a workaround to use ldapwhoami with the IP address
of the domain controller instead of the domain name, taking out the need
to perform name resolution.  With this patch, dc-cac-ws finishes in
approx. 19 minutes.

Signed-off-by: Sherman Yin <syin@teradici.com>