Workstations to get startup scripts from bucket
Instead of logging into workstations to upload and execute startup
scripts, upload the scripts to a bucket and refer to those scripts from
the workstations.

This makes it possible for workstations to be created without a public
IP in the future, and secrets are no longer in the meta-data field of
the instances.

However, workstation provisioning process can no longer be seen from
terraform host.

Signed-off-by: Sherman Yin <syin@teradici.com>