Use CAM service account instead of CAC token (multi-region)
To support autoscaling of CAC's in the future, we will need to generate
new CAC tokens to be used during CAC installations. In order to generate
CAC tokens, we have used a python script which will make calls to the
CAM API using the user-provided CAM service account credentials.

Both plaintext and encrypted secrets is supported for this multi-region
deployment using CAM service account instead of CAC token.

Signed-off-by: Edwin Pau <epau@teradici.com>