Automate the creation of tfvars with encrypted secrets
The python script that encrypts the secrets using the user provided
GCP service account will now create a a new terraform.tfvars file
with the encrypted secrets rendered into the appropriate fields.
The script overwrites the existing tfvars file with plaintext secrets,
and the result is a terraform.tfvars file that is ready to be used for
deployments (both single and multi-region are supported).

Signed-off-by: Edwin Pau <epau@teradici.com>