1/2 AWS: use vars for PCoIP Agent location
As part of the startup scripts for CentOS machines, the download
locations for PCoIP agent has been changed to use terraform variables
for better flexibility and to remain consistent with the Windows
startup scripts. Windows startup scripts were modified to match the
style of the Centos startup scripts using global variables.

Signed-off-by: Edwin Pau <epau@teradici.com>