Update terraform.tfvars samples for GCP
Sample files for terraform.tfvars were modified to ensure that secrets
can be detected by Tfvars_Encryptor_GCP after a specified line in the
file. All secrets below the line "# <-- Start of secrets section,
do not edit this line. -->" will be encrypted by the script. The default
for the kms_cryptokey_id line has also been changed to a comment.

Signed-off-by: Edwin Pau <epau@teradici.com>