GCP: fix quickstart deployment (write tfvars)
A recent commit from encryption/decryption script changed the default
of the kms_cryptokey_id line inside of terraform.tfvars to be a comment
instead. This caused quickstart deployment scripts to not use encryption
since it did not detect a kms_cryptokey_id variable, which caused
the deployment to fail.

The fix explicitly checks for this line and adds it to the quickstart
script first so that the terraform.tfvars is properly created.

Signed-off-by: Edwin-Pau <epau@teradici.com>