GCP: change log filenames
Log filenames have been changed to use "provisioning.log" so that it is
consistent across all VM's.

Fixed CAC logging as it was missing a -a flag, which meant the log was
being overwritten by the CAC installer.

There are now two CAC logs: one for VM provisioning like other VM's,
and another cac-install.log for the CAC installer.

Signed-off-by: Edwin-Pau <epau@teradici.com>