AWS: Use allowable AZs (lb-connectors)
A data source for availability zones has been added so that we can
check and use the available zones during a deployment. A blacklist of
availability zones has also been added to avoid zones that do not have
the required EC2 instances.

The DC and WS subnets have been configured to use the first available
and non-blacklisted availability zone in the region.

Signed-off-by: Edwin-Pau <epau@teradici.com>