Add AWS support to KMS encryption script
-This is an initial commit for the combined kms_secrets_encryption script
which will allow the encryption and decryption of both GCP and AWS
terraform.tfvars files.

-This refactored script will contain four classes: Tfvars_Parser class
which is used by the abstract class Tfvars_Encryptor. Concrete sub
classes GCP_Tfvars_Encryptor and AWS_Tfvars_Encryptor will inherit from
the Tfvars_Encryptor abstract class.

-Rearranged attributes and methods so that the common ones can take
advantage of polymorphism and inheritance based on the encryptor that is
created.

-Modified two of the Tfvars_Encryptor methods to allow for polymorphic
usage which is dependent on the type of instance the encryptor is at
run time. If the instance is a GCP_Tfvars_Encryptor, it uses
"kms_cryptokey_id" for the crypto key. If the instance is a
AWS_Tfvars_Encryptor, it uses "customer_master_key_id" instead.

-Added a function import_or_install_module(pip_package, module_name)
which attempts to import a specified Python package, and if it fails
it will attempt to install it first before importing it again.

-Replaced all usage of str.format() with f-strings instead in the
KMS encryption script.

-Moved error strings to SystemExit() calls instead of using print
statements.

-Added missing try and exception blocks when attempting to open files,
so that any file errors can be caught in the correct code blocks.

-Removed 'r' mode since this is the default mode.

-Changed the 'path' variable name to be 'file_path' so that it's
consistent with other methods.

-Added an explicit check for the KMS encryption script when reading
a terraform.tfvars file for secrets. If the start of secrets flag is
missing from the terraform.tfvars file, then the script will exit and
inform the user that the start of secrets flag is missing.

-Ported the latest changes from the import fix branch.

Signed-off-by: Edwin-Pau <epau@teradici.com>