GCP: Fix provisioning scripts
For the CentOS provisioning scripts, the logic to check for required
variables are now in a separate function called check_required_vars
just like the CAC provisioning scripts.

For the CAC and CentOS provisioning scripts, the check_required_vars
function has been updated to use double quotes for the "true" string,
representing the true boolean value for consistency.

Fixed logging messages in the above scripts when dealing with missing
required variables as well as error messages.

Moved the installation of ldap_utils to the beginning of the wait_for_dc
function.

Minor grammar fix in a DC script.

Signed-off-by: Edwin-Pau <epau@teradici.com>