AWS: Add function and attributes for reading tfvar
Added a function and the required attributes to read data and secrets
from a user-specified terraform.tfvars file.

The constructor now takes in a path to a user-specified terraform.tfvars
file. Attributes are added for storing the tfvars_data and
tfvars_secrets. Function added for reading and parsing the tfvars file.

Signed-off-by: Edwin-Pau <epau@teradici.com>