Add AWS encryptor class in encryption script
Added the AWS encryptor class which inherits from the abstract encryptor
class. This class contains attributes and methods specific to the AWS
environment when encrypting and decrypting secrets.

Rearranged attributes and methods so that the common ones can take
advantage of polymorphism and inheritance based on the encryptor that is
created.

Moved GCP crypto_key_id credentials variable from the abstract class to
the sub class since AWS uses a different crypto key variable, which is
customer_master_key_id.

Signed-off-by: Edwin-Pau <epau@teradici.com>