Allow polymorphism use for super class methods
Modified two of the Tfvars_Encryptor methods to allow for polymorphism
usage which is dependent on the type of instance the encryptor is at
run time. If the instance is a GCP_Tfvars_Encryptor, it uses
"kms_cryptokey_id" for the crypto key. If the instance is a
AWS_Tfvars_Encryptor, it uses "customer_master_key_id" instead.

Signed-off-by: Edwin-Pau <epau@teradici.com>