GCP: Use local vars and increase retry counter
Retry variables for the CAC scripts have been changed to use local
variables so they are scoped to their respective functions.

Increased the CAC install retry counter to retry up to 10
additional times with one minute in between.

Fixed a minor issue with the CAC install logging message so it only
logs once instead of everytime the retry block runs.

Signed-off-by: Edwin-Pau <epau@teradici.com>