GCP: Fix module import error for encryption script
Added a function import_or_install_module(pip_package, module_name)
which attempts to import a specified Python package, and if it fails
it will attempt to install it first before importing it again.

This function takes in two arguments: the first one is the Python
package name and the second optional parameter is for the import name
(if the import name is different than the package name, e.g.
"google-cloud-kms" for pip3, "google.cloud.kms_v1" for import.

Signed-off-by: Edwin-Pau <epau@teradici.com>