GCP: Use absolute paths in quickstart
Modified the GCP quickstart script to pass in absolute paths instead
of relative paths for the following terraform.tfvars variables:
gcp_credentials_file, cac_admin_ssh_pub_key_file and
centos_admin_ssh_pub_key_file.

This change allows the KMS encryption script to work in both the local
and cloud environments. This is also in-line with the terraform.tfvars
.sample file.

Signed-off-by: Edwin-Pau <epau@teradici.com>