Fix Windows retry function and AWS host renaming
The wrong variable was used when printing out retry attempts, this has
been updated to use the correct variable.

For AWS, renaming of the host is required and would sometimes fail.
'-ErrorAction Stop' has been added to the host renaming command so that
it will catch any general errors and allow the retry function to work
properly.

Signed-off-by: Edwin-Pau <epau@teradici.com>