GCP: Move dependencies from script to instance
The dependencies of several google_storage_bucket_object have been
moved from inside the provisioning script resource to inside the CAC
instance resource instead.

The dependency for the provisioning script has also been commented out,
since this should be inferred by Terraform when creating the instance.

Renamed the resource for the CAM credentials file to be
cam-deployment-sa-file so that it's the same for both cac and cac-igm.

Signed-off-by: Edwin-Pau <epau@teradici.com>