GCP: Ensure requirements met for quickstart
Added two functions which ensure the requirements are met before
proceeding with the quickstart deployment.

Moved all the logic to check and install Terraform into the new
function renamed ensure_terraform(). Moved all the logic to check
and install Google libraries into the new function renamed
ensure_google_libraries().

If they are not installed, the script prompts the user
for permission to install the requirements for both Terraform and
Google libraries.

Added Terraform version checking.

Fixed the correct Python packages to check for and install.

Signed-off-by: Edwin-Pau <epau@teradici.com>