GCP: Add Google packages version checking
The most recent version of the google-cloud-kms library v.2.0.0
released on July 31, 2020 includes substantial interface changes.

Existing code written for earlier versions of this library require
updates to use this version. This document describes the changes
that have been made.
https://googleapis.dev/python/cloudkms/latest/UPGRADING.html

The GCP quickstart script has been updated to also check for updated
Google packages. If the package is not >= to the specified version,
then it will prompt the user to update the required packages before
proceeding.

Signed-off-by: Edwin-Pau <epau@teradici.com>