Change the endpoint used to get IP address
The data resource to get the Terraform host's IP address has been
changed to use ifconfig.co/ip instead of ipinfo.io/ip.

This change was made to address a problem with the ipinfo.io/ip endpoint
since if Terraform is deployed from a GCP instance, it would return an
incorrect IP of 0.0.0.0. The endpoint was tested using the command
'curl -v ipinfo.io/ip' and returns a status code of 200 even if it
failed to get the correct IP.

StatusCode        : 200
StatusDescription : OK
Content           : 0.0.0.0

The 0.0.0.0 IP retrieved meant that the firewall rule created by
Terraform would only permit the IP 0.0.0.0/32 to pass through for the
inbound rule, which ultimately caused the Terraform scripts to be stuck
while uploading the provisioning script to the DC using WinRM.

Signed-off-by: Edwin-Pau <epau@teradici.com>