GCP: Split CAC into two modules
The original cac module is a higher level module that creates resources
shared between CACs from different regions.

A cac-regional module is created to contain CACs and region-specific
scripts.

With this method, the CACs are no longer a single list of instances
with different regions. This will help scaling the instances for a
region without affecting the resources from other regions.

This makes use of the new Terrafor 0.13 module features.

Signed-off-by: Sherman Yin <syin@teradici.com>