GCP: Deployment SA file upload at deployment level
As preparation for CAM Standalone support, move the uploading of the CAM
deployment service account key file to the deployment level from the
module level. The CAC module will no longer upload this key itself and
rely on the caller to pre-upload it to the specified storage bucket.
This is needed because in the case of CAM Standalone, the CAM deployment
service account key file will not be located in the Terraform host, but
will be generated as the CAM standalone VM is created.

Signed-off-by: Sherman Yin <syin@teradici.com>