GCP: Update CAC for multi-region deployment
Updated the cac-igm module to use a regional instance group manager
instead of a zonal one, so that CACs are spread across different zones
in the same region for higher availability.

Also updated the cac-igm module to be able to deploye instances without
public IPs and accepting CAC SSL key/cert in prep for using it in the
nlb-multi-region deployment.

Signed-off-by: Sherman Yin <syin@teradici.com>