CentOS: Log NVIDIA Driver Installer Error Message
Currently the code only checks whether the driver installer runs, but
does not check whether the driver installer was downloaded properly or
not (i.e. an empty file would open succesfully and no error is logged).
With this fix, the program will not create a file and will exit upon
error with downloading the installer. To improve the provisioning log's
interpretability, the program will also log the error at each retry.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>