AWS:Fix Centos Graphics Bug
There is a syntax error made in the aws centos graphics provisioning
script that caused erors when the gpu driver was to be installed (There
were comments added after "\" when passing arguments to a method so the
comments were mistaken for arguments instead). The link to the gpu
driver was also updated again because nvidia updated their links again.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>