Check if Idle Shutdown is Installed
The CAM Idle Shutdown installer will give an error if the installer is
executed when the service is already installed. To avoid that, because
an error will cause the script to exit, now the script will first check
whether the service is installed or not and will only execute the
installer if it is not.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>