AWS:Register Service Account Cam Standalone
Register AWS cloud service account to CAM deployment by passing aws
credentials to CAM module and then to the CAM provisioning script. The
user will need to specify a cam_aws_credentials_file in the tfvars that
is a csv file.  Please refer to the cam_aws_credentials.csv.sample file
for the required format.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>