AWS: Fix Key File Decryption Error
Decryption of AWS CAM credentials file resulted in an error because it
was trying to decrypt an empty file. The command created the
output file before reading the file, but since the output file had the
same name as the encrypted file, the encrypted file got replaced with an
empty file. The issue is fixed by renaming the output file.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>