Revert "AWS: Update Documentation For KMS Auto Encryption"
This commit was added to ensure that the key used by the
kms_secrets_encryption.py script is from the same region that was
specified in the tfvars, by relying on the user to run aws configure.
However, a better solution has been proposed that will not rely on the
user.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>