AWS: Pass Region Variable To Provisioning Scripts
The KMS decryptor relies on the correct region to be specified, i.e. the
region which the key exists in. The aws_region specified in the
tfvars is now passed to the provisioning scripts for centos and windows
machines so it will work for any region, not just the default.

Signed-off-by: Nathania Hendradjaja nhendradjaja@teradici.com