AWS: Filter Availability Zone
In the case where a user is using an account that is opted-in to the AWS
Wavelength Zone services, the availability zone assigned to the subnets
might be a wavelength zone instead, which would cause an error when
creating the Route53 resolver and creating the workstations, because
some instace types are not offered in wavelength zones. A filter is
added to assign only availability zones types.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>