AWS: Dependencies On Domain Controller Private IP Address
When a resolver rule association is created, AWS automatically creates a
network interface for the association and assigns it some private IP
address in the VPC IP range, which may be the private IP that was
reserved for the Domain Controller. When this happens, the Terraform
deployment will run into an error saying that the private IP is already
in use. The fix is to let AWS decide on the private IP and make the
Route53 resolver rule depend on and be created after the Domain
Controller instance.

Signed-off-by: Nathania Hendradjaja <nhendradjaja@teradici.com>