AWS: create a dashboard for CAS Connector
It could be difficult for users to monitor the cas connector by reading
logs that contains a variety of messages. To help users to monitor the
connector, we create a dashboard that contains graphs for monitoring
number of users and machines in AD, number of active sessions, and user
login time and workstation.

Signed-off-by: Qi Dang <qidang@hp.com>