RM67893 Allow keypair and credentials as vars
The existing code creates a keypair using a public key file read from
disk, and also uploads aws credentials to S3 by reading a file from
disk.  While this can work from the command line, our standard model
of operation is to upload inventories to Terraform Cloud and apply
the deployment from there.  The Terraform Cloud deployments will fail
if we try to use the file-based approach.

To work around this, we introduce changes to
1. Allow the keypair name to be passed in; if this is done then the
code skips creating a keypair since it won't be needed
2. Allow the credentials file _contents_ to be passed in as an
encrypted string.  When this is done the code creates a _local_ file
with those contents, and then the code proceeds as normal to upload
the contents of the now _local_ file to s3.

References RM-67893