Updated Readme.md and Added CustomRole.json in GCP Docs
The provided group of GCP managed roles outlines the required
permissions for Terraform deployments. In addition, a custom role is
provided as an alternative, which imposes stricter restrictions
compared to the group of GCP managed roles, ensuring limited access. By
utilizing the custom role, the security risk is minimized by granting
only the necessary privileges.

Signed-off-by: Pvs <manogna.pvs@hp.com>